+++ src/ReactDashboard.js
@@ -1,5 +1,4 @@
-import './stylesheets/main.scss';
-console.log('React Dashboard -- 0.6.13');
+console.log('React Dashboard -- 0.6.14.2.1');
 
 // COMPONENTS
 export {default as Autocomplete } from './components/Autocomplete';